server/requirements.txt
fastapi==0.117.1
uvicorn==0.37.0
sqlalchemy==1.4.49
pyjwt[crypto]==2.10.1
pydantic==2.11.9
pydantic-settings==2.3.4
python-multipart==0.0.20
httpx==0.27.2
chess>=1.10.0
xgboost>=2.0.0
numpy>=1.24.0
joblib>=1.3.0
